+++ modules/tinymce/src/core/main/ts/api/Options.ts
@@ -5,7 +5,7 @@ import * as Pattern from '../textpatterns/core/Pattern';
 import * as PatternTypes from '../textpatterns/core/PatternTypes';
 import DOMUtils from './dom/DOMUtils';
 import Editor from './Editor';
-import { EditorOptions } from './OptionTypes';
+import { EditorOptions, ForceHexColor } from './OptionTypes';
 import I18n from './util/I18n';
 import Tools from './util/Tools';
 
@@ -827,6 +827,15 @@ const register = (editor: Editor): void => {
     default: ''
   });
 
+  registerOption('force_hex_color', {
+    processor: (value) => {
+      const options: ForceHexColor[] = [ 'always', 'rgb_only', 'off' ];
+      const valid = Arr.contains(options, value);
+      return valid ? { value, valid } : { valid: false, message: `Must be one of: ${options.join(', ')}.` };
+    },
+    default: 'off',
+  });
+
   // These options must be registered later in the init sequence due to their default values
   editor.on('ScriptsLoaded', () => {
     registerOption('directionality', {
@@ -956,6 +965,8 @@ const hasTableTabNavigation = option('table_tab_navigation');
 const getDetailsInitialState = option('details_initial_state');
 const getDetailsSerializedState = option('details_serialized_state');
 
+const shouldForceHexColor = option('force_hex_color');
+
 export {
   register,
 
@@ -1059,5 +1070,6 @@ export {
   shouldSanitizeXss,
   getDetailsInitialState,
   getDetailsSerializedState,
-  shouldUseDocumentWrite
+  shouldUseDocumentWrite,
+  shouldForceHexColor
 };